Move Cassandra session implementations into a separate file.
